requirements.txt
requests>=2.28.0
pypdf>=4.0.0
PySide6>=6.4.0
pyinstaller>=5.0.0
beautifulsoup4>=4.0.0
Pillow>=10.0.0
